fix kex compilation errors
